backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.0
pdfplumber==0.10.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
python-dotenv==1.0.0
cors==1.0.1
pandas==2.0.3
numpy==1.24.3
# Optional: Uncomment to enable R workload predictor
# rpy2==3.17.0
